Please check my 3-d homework. Thanks.
